+++ src/controllers/WebhookController.php
@@ -9,22 +9,25 @@
 namespace enupal\backup\controllers;
 
 use Craft;
+use craft\helpers\UrlHelper;
 use craft\web\Controller as BaseController;
 
 use enupal\backup\Backup;
 
 class WebhookController extends BaseController
 {
-    protected $allowAnonymous = ['actionFinished', 'actionSchedule'];
+    protected $allowAnonymous = ['actionFinished', 'actionSchedule', 'actionGoogleDriveAuth'];
+
+    // Disable CSRF validation for the entire controller
+    public $enableCsrfValidation = false;
+
 
     /**
      * WebHook to listen when a backup process finish up
      *
-     * @return \yii\web\Response
-     * @throws \Exception
+     * @return mixed
      * @throws \Throwable
-     * @throws \yii\base\Exception
-     * @throws \yii\db\Exception
+     * @throws \craft\web\twig\TemplateLoaderException
      */
     public function actionFinished()
     {
@@ -81,5 +84,45 @@ class WebhookController extends BaseController
         }
 
         return $this->asJson($response);
+    }
+
+    /**
+     * @return \yii\web\Response
+     * @throws \yii\base\Exception
+     */
+    public function actionGoogleDriveAuth()
+    {
+        $key = Craft::$app->request->getParam('code');
+        $response = [
+            'success' => false
+        ];
+
+        if ($key) {
+            $client = Backup::$app->settings->createAccessClient();
+            $accessToken = $client->fetchAccessTokenWithAuthCode($key);
+
+            if (!isset($accessToken['error_description'])){
+                $accessFile = Backup::$app->backups->getGoogleDriveAccessPath();
+
+                // Store the credentials to disk.
+                $basePath = Backup::$app->backups->getBasePath();
+                if (!file_exists($basePath)) {
+                    mkdir($basePath, 0777, true);
+                }
+
+                file_put_contents($accessFile, json_encode($accessToken));
+
+                Craft::$app->getSession()->setNotice(Backup::t('Google Drive successfully added'));
+
+            }else{
+                Craft::$app->getSession()->setError(Backup::t('Couldn’t save google drive access token: '.$accessToken['error_description']));
+            }
+        } else {
+            Craft::$app->getSession()->setError(Backup::t('Empty code response from google drive'));
+        }
+
+        Craft::$app->getView()->setTemplateMode('cp');
+
+        return $this->renderTemplate('enupal-backup/settings/googledrive');
     }
 }