+++ src/elements/db/BackupQuery.php
@@ -91,6 +91,7 @@ class BackupQuery extends ElementQuery
             'enupalbackup_backups.rsync',
             'enupalbackup_backups.ftp',
             'enupalbackup_backups.softlayer',
+            'enupalbackup_backups.googleDrive',
             'enupalbackup_backups.logMessage',
             'enupalbackup_backups.isEncrypted',
         ]);